cli/connectivity: tolerate exec-proxy blips while waiting for NodePorts
WaitForNodePorts probes every NodePort on every node by exec'ing "nc -z"
into a client pod through the kube-apiserver to kubelet exec proxy. On
managed clusters (seen on AKS) that proxy is occasionally unreachable
for tens of seconds at a time, longer than the ShortTimeout the wait
runs under. When that happens the exec fails with, for example, "dial
tcp <apiserver>:443: i/o timeout" rather than a result from nc, and the
old loop counted those proxy failures against the same 30s readiness
budget as a genuinely-not-ready NodePort. A single ~30s proxy blip
during the concurrent NodePort probing (test-concurrency=5 fires many
execs at once) therefore exhausted the budget and failed the whole
connectivity test during setup, with no real test failure.

Separate the two failure modes. A transient exec-proxy error is not a
verdict on whether the NodePort is up, so it now gets its own generous
allowance and the exec runs against the parent context instead of the
ShortTimeout one, so a proxy outage no longer eats into the readiness
budget. Any other error still means nc actually ran and the NodePort is
not ready yet, so it keeps counting against the ShortTimeout readiness
deadline exactly as before.

The classification of a transient exec-proxy error is the same one cilium
features status already needs, so move it next to ExecInPod as the
exported k8s.IsTransientExecError and have both callers use it, adding
the "i/o timeout" fragment for the dial failure observed here. This keeps
the single list of proxy-tunnel error fragments in one place, in the
package that owns the exec call that produces them.

This commit was prepared with AIL:3.
